backend/requirements.txt
fastapi
uvicorn
gunicorn
firebase-admin
python-jose[cryptography]
python-multipart
requests
google-cloud-storage
google-cloud-firestore
pydantic